piston: Move draw_tents and animate_tents to ShooterState methods
We finally made our mind to make references Rc in TempEntity to avoid
lifetime annotation, because wasm-bindgen can't make a struct exposed
to Wasm runtime if it has lifetime.

It is unfortunate but Rc shouldn't cost much since most of the images are
static, the reference counts would not change much. Of course TempEntity
is an exception, but that's a cost we can't afford to avoid.